Bug fix with integer divisions in the Kaiser windows
git-svn-id: file:///home/roberto/Git/svn@2183 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: f77eb628ce80b4dfa2d312ee45e1035474636ff8

+++ trunk/xmipp/libraries/data/mask.cpp
@@ -72,7 +72,7 @@ void KaiserMask(Matrix1D<double> &mask, double delta, double Deltaw)
 
     // Design Kaiser window
     double A = -20 * log10(delta);
-    int    M = CEIL((A - 8) / (2.285 * Deltaw));
+    double M = CEIL((A - 8) / (2.285 * Deltaw));
     double beta;
     if (A > 50)
         beta = 0.1102 * (A - 8.7);
@@ -380,7 +380,7 @@ void SeparableSincKaiserMask(Matrix2D<double> &mask,
 
     // Design Kaiser window
     double A = -20 * log10(delta);
-    int    M = CEIL((A - 8) / (2.285 * Deltaw));
+    double M = CEIL((A - 8) / (2.285 * Deltaw));
     double beta;
     if (A > 50)
         beta = 0.1102 * (A - 8.7);
